增加cuda扩展接口

+++ src/main/java/org/dromara/easyai/matrixTools/CudaMatrix.java
@@ -9,14 +9,14 @@ public interface CudaMatrix {
 	Matrix mulMatrix(Matrix matrix1, Matrix matrix2) throws Exception;
 
 	// 矩阵数加
-	void mathAdd(Matrix matrix, double nub) throws Exception;
+	void mathAdd(Matrix matrix, float nub) throws Exception;
 
 	// 矩阵数减
-	void mathSub(Matrix matrix, double nub) throws Exception;
+	void mathSub(Matrix matrix, float nub) throws Exception;
 
 	// 矩阵数乘
-	void mathMul(Matrix matrix, double nub) throws Exception;
+	void mathMul(Matrix matrix, float nub) throws Exception;
 
 	// 矩阵数除
-	void mathDiv(Matrix matrix, double nub) throws Exception;
+	void mathDiv(Matrix matrix, float nub) throws Exception;
 }
